+++ docs/assets/mdx-react-shim-CDJ2MSQt.js
@@ -1,1 +1,0 @@
-import{n as e}from"./chunk-zsgVPwQN.js";import{t}from"./react-my-CdR7O.js";var n=e((()=>{t()}));export{n as t};