+++ docs/assets/mdx-react-shim-N0-qc3-3.js
@@ -1,0 +1,1 @@
+import{n as e}from"./chunk-zsgVPwQN.js";import{t}from"./react-D53p80z6.js";var n=e((()=>{t()}));export{n as t};