+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This is an alpha version! The changes listed here are not final.
 
+### Added
+- Add premium-analytics internal packages stories
+
 ### Changed
 - Internal: No longer require automattic/jetpack-changelogger as a per-project dev dependency.
 - Storybook Vite plugin: resolve bare `store` and `store/` imports for search dashboard stories.